[REAFACTOR] Refactor File Structure
