+++ app/src/main/java/com/example/pocky/ComboActivity.java
@@ -7,7 +7,7 @@ import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatButton;
 
-import com.example.pocky.mainActivity.MainActivity;
+import com.example.pocky.presentation.screen.main.MainActivity;
 
 public class ComboActivity extends AppCompatActivity {
