+++ app/src/main/java/com/example/pocky/domain/model/favor/FavorInterface.java
@@ -1,0 +1,20 @@
+package com.example.pocky.domain.model.favor;
+
+import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.http.Body;
+import retrofit2.http.GET;
+import retrofit2.http.Headers;
+import retrofit2.http.POST;
+import retrofit2.http.Query;
+
+public interface FavorInterface {
+
+    @GET("/favor/getfavor")
+    Call<List<FavorDTO>> getFavor(@Query("id") String id);
+
+    @POST("/favor/postfavor")
+    @Headers("Content-type: application/json")
+    Call<FavorDTO> setFavorInfo(@Body FavorDTO favordto);
+}
