+++ app/src/main/java/com/example/pocky/domain/usecase/QrCodeActivity.java
@@ -1,0 +1,30 @@
+package com.example.pocky.domain.usecase;
+
+import android.graphics.Bitmap;
+import android.os.Bundle;
+import android.widget.ImageView;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.pocky.R;
+import com.google.zxing.BarcodeFormat;
+import com.google.zxing.WriterException;
+import com.journeyapps.barcodescanner.BarcodeEncoder;
+
+public class QrCodeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_qr);
+
+        try {
+            BarcodeEncoder barcodeEncoder = new BarcodeEncoder();
+            Bitmap bitmap = barcodeEncoder.encodeBitmap("Hello, world!", BarcodeFormat.QR_CODE, 400, 400);
+            ImageView imageViewQrcode = findViewById(R.id.imageViewQrCode);
+            imageViewQrcode.setImageBitmap(bitmap);
+        } catch (WriterException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
