--- app/src/main/java/com/example/pocky/model/user/UserDTO.java
@@ -1,35 +1,0 @@
-package com.example.pocky.model.user;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class UserDTO {
-    @SerializedName("id")
-    @Expose
-    private String id;
-
-    @SerializedName("Nickname")
-    @Expose
-    private String NickName;
-
-    public UserDTO(String id, String nickName) {
-        this.id = id;
-        this.NickName = nickName;
-    }
-    public String getNickName() {
-        return NickName;
-    }
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public void setNickName(String nickName) {
-        NickName = nickName;
-    }
-
-
-}
