--- app/src/main/java/com/example/pocky/moreActivity/MoreActivity.java
@@ -1,15 +1,0 @@
-package com.example.pocky.moreActivity;
-
-import android.os.Bundle;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-public class MoreActivity extends AppCompatActivity {
-
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-}
