+++ app/src/main/java/com/example/pocky/presentation/screen/SplashActivity.java
@@ -1,0 +1,36 @@
+package com.example.pocky.presentation.screen;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.pocky.databinding.ActivitySplashBinding;
+import com.example.pocky.presentation.screen.loginActivity.LoginActivity;
+
+public class SplashActivity extends AppCompatActivity {
+
+    private ActivitySplashBinding binding;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        binding = ActivitySplashBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+
+        Handler handler = new Handler(Looper.getMainLooper());
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                // 일정 시간이 지나면 MainActivity로 이동
+                Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
+                startActivity(intent);
+
+                // 이전 키를 눌렀을 때 스플래스 스크린 화면으로 이동을 방지하기 위해
+                // 이동한 다음 사용안함으로 finish 처리
+                finish();
+            }
+        }, 2000); // 시간 2초 이후 실행
+    }
+}
