+++ app/src/main/java/com/example/pocky/presentation/screen/main/frgment/more/MoreActivity.java
@@ -1,0 +1,15 @@
+package com.example.pocky.presentation.screen.main.frgment.more;
+
+import android.os.Bundle;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class MoreActivity extends AppCompatActivity {
+
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+}
