+++ app/src/main/java/com/example/pocky/presentation/screen/shoppingCartActivity/ShoppingCartActivity.java
@@ -1,0 +1,13 @@
+package com.example.pocky.presentation.screen.shoppingCartActivity;
+
+import android.os.Bundle;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class ShoppingCartActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+}
