+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="item"
-            type="com.example.pocky.mainActivity.MainActivity" />
+            type="com.example.pocky.presentation.screen.main.MainActivity" />
         <variable
             name="name"
             type="String" />
